prettify vm tests
